Move validateContractsAndConvert to helper file.
